pyproject.toml
[project]
name = "bird-survey"
version = "0.1.1"
description = "AI-powered bird survey data extraction tool"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.15.0",
    "itrx>=0.1.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pandas-stubs>=2.2.3.250308",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
    "typer>=0.15.4",
    "xlsxwriter>=3.2.3",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.11.10",
    "types-openpyxl>=3.1.5.20250516",
]

[tool.uv]
package = false

[tool.ruff.lint]
select = ["E", "F", "W", "Q", "I", "UP", "B", "SIM"]

[tool.ruff]
line-length = 120
